Move Effects and Matchers to their own files
